Factor out CeleryScript compiler into multiple modules
This should hopefully allow for easier to follow error messages as it
removes all elixir macro "magic"
